feat: add husky script
