+++ .github/workflows/deploy.yml
@@ -10,6 +10,9 @@ on:
 
   workflow_dispatch:
 
+env:
+  HUSKY: 0
+
 permissions:
   contents: read
   pages: write
