+++ .husky/pre-push
@@ -1,0 +1,1 @@
+pnpm lint && pnpm coverage && pnpm type-check
