+++ package.json
@@ -11,7 +11,8 @@
     "type-check": "tsc --project ./tsconfig.json --noEmit --skipLibCheck",
     "lint": "eslint --cache .",
     "test": "vitest",
-    "coverage": "vitest run --coverage"
+    "coverage": "vitest run --coverage",
+    "prepare": "husky"
   },
   "dependencies": {
     "react": "^18.3.1",
@@ -34,6 +35,7 @@
     "eslint-plugin-react": "^7.34.2",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.7",
+    "husky": "^9.0.11",
     "jsdom": "^24.1.0",
     "prettier": "^3.3.2",
     "typescript": "^5.4.5",
